admiralonco: Oncology Extension Package for ADaM in 'R' Asset Library

Programming oncology specific Clinical Data Interchange Standards Consortium (CDISC) compliant Analysis Data Model (ADaM) datasets in 'R'. ADaM datasets are a mandatory part of any New Drug or Biologics License Application submitted to the United States Food and Drug Administration (FDA). Analysis derivations are implemented in accordance with the "Analysis Data Model Implementation Guide" (CDISC Analysis Data Model Team (2021), https://www.cdisc.org/standards/foundational/adam). The package is an extension package of the 'admiral' package.

See also

Author

Maintainer: Stefan Bundfuss stefan.bundfuss@roche.com

Authors:

Amit Jain

Vinh Nguyen

Olga Starostecka

Kiran Peddamudium

Tomoyuki Namai

Ross Farrugia

Other contributors:

Yirong Cao [contributor]

Ashwini Weber [contributor]

F. Hoffmann-La Roche AG [copyright holder, funder]

GlaxoSmithKline LLC [copyright holder, funder]

Amgen Inc. [copyright holder, funder]

Bristol Myers Squibb [copyright holder, funder]