Pre-Defined Time-to-Event Source Objects

These pre-defined tte_source objects can be used as input to 'admiral::derive_param_tte()`.

death_event

lastalive_censor

pd_event

lasta_censor

rand_censor

trts_censor

Format

An object of class censor_source (inherits from tte_source, source, list) of length 5.

An object of class event_source (inherits from tte_source, source, list) of length 5.

An object of class censor_source (inherits from tte_source, source, list) of length 5.

An object of class censor_source (inherits from tte_source, source, list) of length 5.

An object of class censor_source (inherits from tte_source, source, list) of length 5.

Details

To see the definition of the various objects simply print the object in the R console, e.g. print(death_event). For details of how to use these objects please refer to admiral::derive_param_tte().

Printing an object will display input dataset_name, filter (if applicable), date variable, and appropriate values for EVNTDESC, CNSDTDSC, SRCDOM, SRCVAR, and SRCSEQ.

See also

Examples

# This shows the definition of all pre-defined `tte_source` objects that ship
# with {admiralonco}
for (obj in admiral::list_tte_source_objects(package = "admiralonco")$object) {
  cat(obj, "\n")
  print(get(obj))
  cat("\n")
}
#> lasta_censor 
#> <tte_source> object
#> dataset_name: "adrs"
#> filter: PARAMCD == "LSTA" & ANL01FL == "Y"
#> date: ADT
#> censor: 1
#> set_values_to:
#>   EVNTDESC: "Last Tumor Assessment"
#>   CNSDTDSC: "Last Tumor Assessment"
#>   SRCDOM: "ADRS"
#>   SRCVAR: "ADT"
#>   SRCSEQ: ASEQ
#> 
#> trts_censor 
#> <tte_source> object
#> dataset_name: "adsl"
#> filter: NULL
#> date: TRTSDT
#> censor: 1
#> set_values_to:
#>   EVNTDESC: "Treatment Start"
#>   CNSDTDSC: "Treatment Start"
#>   SRCDOM: "ADSL"
#>   SRCVAR: "TRTSDT"
#> 
#> rand_censor 
#> <tte_source> object
#> dataset_name: "adsl"
#> filter: NULL
#> date: RANDDT
#> censor: 1
#> set_values_to:
#>   EVNTDESC: "Randomization"
#>   CNSDTDSC: "Randomization"
#>   SRCDOM: "ADSL"
#>   SRCVAR: "RANDDT"
#> 
#> lastalive_censor 
#> <tte_source> object
#> dataset_name: "adsl"
#> filter: NULL
#> date: LSTALVDT
#> censor: 1
#> set_values_to:
#>   EVNTDESC: "ALIVE"
#>   SRCDOM: "ADSL"
#>   SRCVAR: "LSTALVDT"
#> 
#> death_event 
#> <tte_source> object
#> dataset_name: "adsl"
#> filter: DTHFL == "Y"
#> date: DTHDT
#> censor: 0
#> set_values_to:
#>   EVNTDESC: "DEATH"
#>   SRCDOM: "ADSL"
#>   SRCVAR: "DTHDT"
#> 
#> pd_event 
#> <tte_source> object
#> dataset_name: "adrs"
#> filter: PARAMCD == "PD" & AVALC == "Y" & ANL01FL == "Y"
#> date: ADT
#> censor: 0
#> set_values_to:
#>   EVNTDESC: "Disease Progression"
#>   SRCDOM: "ADRS"
#>   SRCVAR: "ADT"
#>   SRCSEQ: ASEQ
#>