admiralonco: Oncology Extension Package for ADaM in 'R' Asset Library

Programming oncology specific Clinical Data Interchange Standards Consortium (CDISC) compliant Analysis Data Model (ADaM) datasets in 'R'. ADaM datasets are a mandatory part of any New Drug or Biologics License Application submitted to the United States Food and Drug Administration (FDA). Analysis derivations are implemented in accordance with the "Analysis Data Model Implementation Guide" (CDISC Analysis Data Model Team (2021), https://www.cdisc.org/standards/foundational/adam). The package is an extension package of the 'admiral' package.

See also

Author

Maintainer: Stefan Bundfuss stefan.bundfuss@roche.com

Authors:

Ross Farrugia

Amit Jain

Andrew Smith

Teckla Akinyi

Samia Kabi

Stephen Gormley

Hinal Patel

Vinh Nguyen

Olga Starostecka

Other contributors:

Uwe Bader [contributor]

Karl Kennedy [contributor]

Edoardo Mancini [contributor]

Matt Marino [contributor]

Gopi Vegesna [contributor]

Thomas Neitmann [contributor]

Annie Yang [contributor]

Konstantina Koukourikou [contributor]

Pavan Kumar [contributor]

Liming Jin [contributor]

Yirong Cao [contributor]

Zhang Kangjie [contributor]

F. Hoffmann-La Roche AG [copyright holder, funder]

GlaxoSmithKline LLC [copyright holder, funder]

Amgen Inc. [copyright holder, funder]

Bristol Myers Squibb [copyright holder, funder]